requirements.txt
fastapi
uvicorn
requests
python-dotenv
pymongo
pydantic
python-multipart
PyMuPDF
easyocr
paddleocr
paddlepaddle
opencv-python
langchain
langchain-community
langchain-core
langchain-huggingface
pytesseract
pillow
gTTS
